+++ app/__init__.py
@@ -34,10 +34,11 @@ def not_found(error):
 
 # Import a module / component using its blueprint handler variable (mod_auth)
 from app.mod_auth.controllers import mod_auth as auth_module
+from app.mod_main.controllers import mod_main as main_module
 
 # Register blueprint(s)
 app.register_blueprint(auth_module)
-# app.register_blueprint(xyz_module)
+app.register_blueprint(main_module)
 # ..
 
 # Build the database:
