+++ app/mod_main/models.py
@@ -15,7 +15,6 @@ class Base(db.Model):
 
     __abstract__ = True
 
-
     date_created = db.Column(db.DateTime,  default=db.func.current_timestamp())
     date_modified = db.Column(db.DateTime,  default=db.func.current_timestamp(), onupdate=db.func.current_timestamp())
