+++ app/templates/auth/login.html
@@ -1,4 +1,4 @@
-{% extends "auth/base.html" %}
+{% extends "main/base.html" %}
 
 {% block content %}
     <div class="row-fluid">
