--- app/templates/auth/viewentry.html
@@ -1,54 +1,0 @@
-{% extends "auth/base.html" %}
-{% block content %}
-
-    <div class="row-fluid">
-            <div class="col-lg-3">
-            </div>
-            <div id="center-side-id4" class="col-lg-6 center-side">
-                {% block center_content %}
-                    {% if journal_entry %}
-                        <div class="panel panel-primary">
-                            <div class="panel-heading"><h3><b>{{ journal_entry.title|upper }}</b></h3></div>
-                            <div id="panel-light" class="panel-body">{{ journal_entry.body }}</div>
-                            <div class="panel-footer"><i><small>{{ journal_entry.tags }}</small></i></div>
-                        </div>
-                        <div class="btn-group btn-group-justified">
-                            <a class="btn btn-success" href="{{ url_for('auth.update_entry', id = journal_entry.id) }}"> Edit </a>
-                            <a class="btn btn-danger" href="{{ url_for('auth.delete', id=journal_entry.id) }}"> Delete</a>
-{#                           <a class="btn btn-danger" href="#"> Delete</a>#}
-                        </div>
-                    {% else %}
-                    {% endif %}
-                {% endblock %}
-            </div>
-
-            <div class="col-lg-3">
-                {% block right_content %}
-                    {{ super() }}
-                {% endblock %}
-            </div>
-        </div>
-
-{#    <div>#}
-{#        {% if journal_entry %}#}
-{#            <div style="text-align: center; margin-bottom: 20px" >#}
-{#                <h3 style="margin-bottom: 10px"> Title: {{ journal_entry.title }}</h3>#}
-{#                <h4> Author: </h4>#}
-{#                <p> {{ user.username}} </p>#}
-{#                <h4> Body: </h4>#}
-{#                <p> {{ journal_entry.body }} </p>#}
-{#                <h4> Tags: </h4>#}
-{#                <p> {{ journal_entry.tags }} </p>#}
-{#              </div>#}
-{##}
-
-{##}
-{#      {% else %}#}
-{#          <div style="text-align: center">#}
-{#            <h3> You have no Entries. Click below to add one:</h3>#}
-{#            <a href="{{ url_for('auth.JournalEntry') }}"><button class="button btn btn-primary" style = "margin:0 auto; float: none; margin-top: 20px" >New Entry</button></a>#}
-{#          </div>#}
-{#      {% endif %}#}
-{##}
-{#    </div>#}
-{% endblock %}
