+++ app/templates/main/base.html
@@ -27,10 +27,16 @@
 <body>
 {% block navbar %}
     <nav class="navbar navbar-inverse">
+        <div class="container-fluid">
         <div class="nav navbar-header">
 {#            <a class="navbar-brand" href="{{ url_for('auth.index') }}">JournalX<em>Memories! Secrets! Privacy!</em></a>#}
-           <a class="navbar-brand" href="{{ url_for('auth.index') }}"> <img src="{{ url_for('static', filename = 'media/logo.png') }}"></a>
-
+           <a class="navbar-brand" href="{{ url_for('main.index') }}"> <img src="{{ url_for('static', filename = 'media/logo.png') }}" alt="JournalX"></a>
+        </div>
+{#        <ul class="nav navbar-bar navbar-right">#}
+{#            {% if g.user.is_authenticated %}#}
+{#                <li><span href="#" >{{ g.user.username }} </span></li>#}
+{#            {% endif %}#}
+{#        </ul>#}
         </div>
     </nav>
 {% endblock %}
@@ -40,7 +46,7 @@
             <div class="col-lg-3 left-side">
                 {% block left_content %}
                     {% if g.user.is_authenticated %}
-                        <a href="{{ url_for('auth.profile') }}" class="btn btn-success btn-group-justified">View Dashboard</a>
+                        <a href="{{ url_for('main.profile') }}" class="btn btn-success btn-group-justified">View Dashboard</a>
                     {% endif %}
                 {% endblock %}
             </div>
@@ -53,7 +59,7 @@
                 {% block right_content %}
                     <div >
                     {% if g.user.is_authenticated %}
-                        <form style="display: inline;" action="{{ url_for('auth.search') }}" method="post"
+                        <form style="display: inline;" action="{{ url_for('main.search') }}" method="post"
                               name="search">
                             {{ g.search_form.hidden_tag() }} {{ g.search_form.search(size=17) }}
                             <input class="btn btn-success btn-sm" type="submit" value="Search">
