+++ app/templates/main/edit.html
@@ -1,4 +1,4 @@
-{% extends "auth/base.html" %}
+{% extends "main/base.html" %}
 
 {% import "bootstrap/wtf.html" as wtf %}
 {% block content %}
