+++ app/templates/main/journalentry.html
@@ -1,4 +1,4 @@
-{% extends "auth/base.html" %}
+{% extends "main/base.html" %}
 
 {% block content %}
     <div class="row-fluid">
@@ -8,7 +8,7 @@
             {% block center_content %}
                 <div class="page-header text-center"><h3><b>New Entry</b></h3></div>
                 <div class="panel panel-body">
-                <form action="{{ url_for('auth.JournalEntry') }}" method=post>
+                <form action="{{ url_for('main.JournalEntry') }}" method=post>
                     <div class="form-group row">
                         <label class="text-left" for="{{ form.title.name }}">
                             {% for message in form.title.errors %}
@@ -32,7 +32,6 @@
 
                     </div>
                 {{ form.submit }}
-{#                    <input class="btn btn-success btn-group-justified" id="submit" name="submit" type="submit" value="Save"#}
                 </form>
                 </div>
             {% endblock %}
